+++ Classes/Progress Bar/TSSTInfoWindow.h
@@ -15,9 +15,9 @@
  */
 @interface TSSTInfoWindow : NSPanel
 
-- (void)caretAtPoint:(NSPoint)point size:(NSSize)size withLimitLeft:(float)left right:(float)right;
+- (void)caretAtPoint:(NSPoint)point size:(NSSize)size withLimitLeft:(CGFloat)left right:(CGFloat)right;
 - (void)centerAtPoint:(NSPoint)center;
-- (void)resizeToDiameter:(float)diameter;
+- (void)resizeToDiameter:(CGFloat)diameter;
 
 @end
 
@@ -26,7 +26,7 @@
 @interface TSSTInfoView : NSView
 
 @property (assign) BOOL bordered;
-@property (nonatomic, assign) float caretPosition;
+@property (nonatomic, assign) CGFloat caretPosition;
 
 @end
